pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "litellm>=1.80.7",
    "openai>=2.8.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "smolagents>=1.23.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
